Add fiber assignement scripts
