+++ acm/hod/cutsky.py
@@ -31,7 +31,7 @@ except ImportError:
     build_healpix_map = None
     HAS_REGRESSIS = False
 
-warnings.filterwarnings("ignore", category=np.exceptions.VisibleDeprecationWarning)
+# warnings.filterwarnings("ignore", category=np.exceptions.VisibleDeprecationWarning)
 LRG_Abacus_DM = get_Abacus_dirs(tracer='LRG', simtype='box')
 
 # Valid DESI photometric regions
@@ -242,7 +242,8 @@ class BaseCutskyCatalog(ABC):
             }
             return pixels, dr9_footprint(convert_dict[region])[pixels]
 
-    def add_columns_fiberassign(self, seed: int = 0) -> None:
+    @staticmethod
+    def add_columns_fiberassign(catalog, seed: int = 0) -> None:
         """
         Add columns to the catalog that are needed for fiber assignment.
 
@@ -251,19 +252,175 @@ class BaseCutskyCatalog(ABC):
         seed : int, optional
             Random seed for reproducibility. Defaults to 0.
         """
-        self.logger.info('Adding columns for fiber assignment.')
 
-        priority = {'LRG': 3200, 'QSO': 3400, 'ELG': 3100, 'BGS': 2100}
-        tile = {'LRG': 'DARK', 'QSO': 'DARK', 'ELG': 'DARK', 'BGS': 'BRIGHT'}
+        priority = {'LRG': 3200, 'QSO': 3400, 'ELG_HIP': 3200, 'ELG_LOP': 3100, 'ELG_VLO': 3000, 'BGS': 2100}
+        numobs = {'LRG': 2, 'QSO': 4, 'ELG_HIP': 2, 'ELG_LOP': 2, 'ELG_VLO': 2, 'BGS': 1}
 
-        csize = len(self.catalog['RA'])
-        self.catalog['DESI_TARGET'] = desi_mask[self.tracer] * np.ones(csize, dtype='i8')
-        self.catalog['PRIORITY'] = priority[self.tracer] * np.ones(csize, dtype='i8')
-        self.catalog['SUBPRIORITY'] = np.random.uniform(size=csize, low=0, high=1).astype('f8')
-        self.catalog['OBSCONDITIONS'] = obsconditions.mask(tile[self.tracer]) * np.ones(csize, dtype='i8')
-        self.catalog['NUMOBS_MORE'] = np.ones(csize, dtype='i8')
-        self.catalog['TARGETID'] = np.arange(csize, dtype='i8')
+        tile = {'LRG': 'DARK', 'QSO': 'DARK', 'ELG': 'DARK', 'ELG_HIP': 'DARK', 'ELG_LOP': 'DARK', 'ELG_VLO': 'DARK', 'BGS': 'BRIGHT'} 
 
+        names_col = ['PRIORITY_INIT', 'PRIORITY', 'NUMOBS_MORE', 'NUMOBS_INIT', 'DESI_TARGET', 'OBSCONDITIONS']
+        for name in names_col:
+            catalog[name] = catalog.ones(dtype=int)
+        
+        # This can be done better using random generator
+        np.random.seed(seed)
+        catalog['TARGETID'] = np.random.permutation(np.arange(1,catalog.size+1)) + catalog.mpicomm.rank * (10**6) # to make iot unique across processes
+        catalog['SUBPRIORITY'] = np.random.uniform(0, 1, catalog.size)
+        tracer_list = np.unique(catalog['TRACER']).tolist()
+        for tr in tracer_list: 
+            mask= catalog['TRACER'] == tr
+            catalog['PRIORITY_INIT'][mask] = priority[tr]
+            catalog['PRIORITY'][mask] = priority[tr]
+            catalog['NUMOBS_MORE'][mask] = numobs[tr]
+            catalog['NUMOBS_INIT'][mask] = numobs[tr]
+            catalog['DESI_TARGET'][mask] = desi_mask[tr]
+            catalog['OBSCONDITIONS'][mask] = obsconditions.mask(tile[tr])
+
+
+
+    def run_FA_for_single_tracer(self, release='Y3', program='dark', npasses=7, use_sky_targets=False, 
+                                 seed=None, preload_sky_targets=False, plot_output=True, mpicomm=None):
+
+        """
+        Run a full fiber assignment workflow on a mock cutsky catalog.
+
+        Optionally adds other tracers randomly, applies multi-pass fiber assignment,
+        computes completeness weights, and produces diagnostic plots.
+
+        Parameters
+        ----------
+        release : str, default='Y3'
+            DESI data release.
+        program : str, default='dark'
+            Observing program.
+        npasses : int, default=7
+            Number of observing passes.
+        use_sky_targets : bool, default=False
+            Include sky targets.
+
+        seed : int or None
+            RNG seed.
+        preload_sky_targets : bool, default=False
+            Preload sky targets.
+        plot_output : bool, default=True
+            Produce diagnostic plots.
+        path_to_save : str or None
+            Output catalog path.
+
+        Returns
+        -------
+        cutsky_for_fa : mpytools.Catalog
+            Catalog with fiber assignment results.
+        """
+        from mpytools import Catalog
+        from mockfactory.desi import build_tiles_for_fa, apply_fiber_assignment, read_sky_targets, compute_completeness_weight
+
+        logger = logging.getLogger('F.A.')
+        rng = np.random.RandomState(seed=seed)
+        
+        if mpicomm is None:
+            from mpi4py import MPI
+            mpicomm = MPI.COMM_WORLD
+        if mpicomm.rank == 0:
+            self.catalog = Catalog.from_dict(self.catalog, mpicomm=mpicomm)
+        self.catalog = Catalog.scatter(self.catalog, mpiroot=0, mpicomm=mpicomm)
+
+        if 'TRACER' not in  self.catalog.columns():
+            logger.info(f'Add tracer column for {self.catalog}.')
+            self.catalog['TRACER'] = [self.tracer]*self.catalog.size
+
+            
+        if mpicomm.rank == 0: 
+            logger.info('Run simple example to illustrate how to run fiber assignment.')
+            logger.info(f'Add random ELGs and QSOs objects.')
+
+        
+        # This is should be done better 
+
+        if 'ELG' in self.tracer:
+            mask_elg_vlo = rng.uniform(size=self.catalog.size) < .25
+            self.catalog['TRACER'][mask_elg_vlo] = 'ELG_VLO'
+            mask_elg_hip = rng.uniform(size=self.catalog.size) < 0.1
+            self.catalog['TRACER'][mask_elg_hip] = 'ELG_HIP'
+            tr_toadd = ['LRG', 'QSO']
+        else:
+            tr_toadd = ['LRG','ELG_HIP', 'QSO']
+            tr_toadd.remove(self.tracer)
+        nbar1 = 240 if tr_toadd[0] == 'ELG_HIP' else 310 if tr_toadd[0] == 'QSO' else 610
+        nbar2 = 240 if tr_toadd[1] == 'ELG_HIP' else 310 if tr_toadd[1] == 'QSO' else 610
+        seed1, seed2 = rng.randint(0,2**32-1,size=2)
+        cutsky_1 = RandomCutskyCatalog(rarange=(self.catalog['RA'].min(), self.catalog['RA'].max()), decrange=(self.catalog['DEC'].min(), self.catalog['DEC'].max()), drange=(self.catalog['Distance'].min(), self.catalog['Distance'].max()), nbar= nbar1, seed=seed1, mpicomm=mpicomm)
+        cutsky_2 = RandomCutskyCatalog(rarange=(self.catalog['RA'].min(), self.catalog['RA'].max()), decrange=(self.catalog['DEC'].min(), self.catalog['DEC'].max()), drange=(self.catalog['Distance'].min(), self.catalog['Distance'].max()), nbar=nbar2, seed=seed2, mpicomm=mpicomm)
+
+        cutsky_1['TRACER'] = [tr_toadd[0]]*cutsky_1.size
+        cutsky_2['TRACER'] = [tr_toadd[1]]*cutsky_2.size
+        cutsky_2 = cutsky_2[is_in_desi_footprint(cutsky_2['RA'], cutsky_2['DEC'], release=release, program=program, npasses=npasses)]
+        cutsky_1 = cutsky_1[is_in_desi_footprint(cutsky_1['RA'], cutsky_1['DEC'], release=release, program=program, npasses=npasses)]
+        cutsky_for_fa = Catalog.concatenate([self.catalog[cutsky_1.columns()], cutsky_1, cutsky_2])
+
+        self.add_columns_fiberassign(cutsky_for_fa)
+        # Collect tiles from surveyops directory on which the fiber assignment will be applied
+        tiles = build_tiles_for_fa(release_tile_path=f'/global/cfs/cdirs/desi/survey/catalogs/{release}/LSS/tiles-{program.upper()}.fits', program=program, npasses=npasses)
+
+        if use_sky_targets and preload_sky_targets:
+            # tiles is not restricted here, we will load sky_targets for all the Y1 footprint
+            sky_targets = read_sky_targets(dirname='/global/cfs/cdirs/desi/users/edmondc/desi_targets/sky_targets/', tiles=tiles, program=program, mpicomm=mpicomm)
+
+        # Get info from origin fiberassign file and setup options for F.A.
+        ts = str(tiles['TILEID'][0]).zfill(6)
+        fht = fitsio.read_header(f'/global/cfs/cdirs/desi/target/fiberassign/tiles/trunk/{ts[:3]}/fiberassign-{ts}.fits.gz')
+        rundate = fht['RUNDATE']
+        # see fiberassign.scripts.assign.parse_assign (Can modify margins, number of sky fibers for each petal etc.)
+        opts_for_fa = ["--target", " ", "--rundate", rundate, "--mask_column", "DESI_TARGET"]
+
+        
+        nbr_targets = cutsky_for_fa.csize
+        if mpicomm.rank == 0: logger.info(f'Keep only objects which is in a tile. Working with {nbr_targets} targets')
+
+        columns_for_fa = ['RA', 'DEC', 'TARGETID', 'DESI_TARGET', 'SUBPRIORITY', 'OBSCONDITIONS', 'NUMOBS_MORE']
+
+        # Let's do the F.A.:
+        apply_fiber_assignment(cutsky_for_fa, tiles, npasses, opts_for_fa, columns_for_fa, mpicomm, use_sky_targets=use_sky_targets)
+        # Compute the completeness weight: if multi-tracer, apply completeness weight once for each tracer independently
+        compute_completeness_weight(cutsky_for_fa, tiles, npasses, mpicomm)
+        # Summarize and plot
+        ra, dec = cutsky_for_fa.cget('RA', mpiroot=0), cutsky_for_fa.cget('DEC', mpiroot=0)
+        numobs, available = cutsky_for_fa.cget('NUMOBS', mpiroot=0), cutsky_for_fa.cget('AVAILABLE', mpiroot=0)
+        obs_pass, comp_weight = cutsky_for_fa.cget('OBS_PASS', mpiroot=0), cutsky_for_fa.cget('COMP_WEIGHT', mpiroot=0)
+
+        logger.info('FA done')
+        
+
+        mask_tr = cutsky_for_fa['TRACER']== self.tracer
+        cutsky_for_fa = cutsky_for_fa[mask_tr]
+        for col in cutsky_for_fa.columns():
+            if col not in self.catalog.columns():
+                self.catalog[col] = cutsky_for_fa[col]
+        del cutsky_for_fa
+        
+        if plot_output & (mpicomm.rank == 0):
+
+            logger.info(f"Nbr of targets observed: {(numobs >= 1).sum()} -- per pass: {obs_pass.sum(axis=0)} -- Nbr of targets available: {available.sum()} -- Nbr of targets: {ra.size}")
+            logger.info(f"In percentage: Observed: {(numobs >= 1).sum()/ra.size:2.2%} -- Available: {available.sum()/ra.size:2.2%}")
+            values, counts = np.unique(comp_weight, return_counts=True)
+            logger.info(f'Sanity check for completeness weight: {available.sum() - (numobs >= 1).sum()} avialable unobserved targets and {np.nansum([(val - 1) * count for val, count in zip(values, counts)])} from completeness counts')
+            logger.info(f'Completeness counts: {values} -- {counts}')
+
+            import skyproj
+            import matplotlib.pyplot as plt
+            fig, ax = plt.subplots(figsize=(12, 12))
+            # sp.draw_des(label='DES', edgecolor='k')
+            sp = skyproj.DESSkyproj(ax=ax, extent=[178,209, 33, 48], fontsize=8)
+
+            sp.scatter(ra[available], dec[available], s=0.001, c='k')
+            sp.scatter(ra[numobs>0], dec[numobs>0], s=0.01,c='r')
+
+            fig.tight_layout()
+            fig.savefig(f'fiberasignment_{npasses}npasses.png', facecolor='w', bbox_inches='tight', pad_inches=0.2, dpi=400)
+            logger.info(f'Plot save in fiberasignment_{npasses}npasses.png')
+        mpicomm.Barrier()
+
+    
     def save(self, filename: str) -> None:
         """
         Save the cutsky catalog to a file. Supports .fits and .npy formats.
@@ -282,7 +439,6 @@ class BaseCutskyCatalog(ABC):
             np.save(filename, self.catalog)
         else:
             raise ValueError('Unsupported file format. Use .fits or .npy.')
-        
 
 class CutskyHOD(BaseCutskyCatalog):
     """
